refactor(network): fix the package name according to guidelines (#1934)
