+++ starport/services/network/publish.go
@@ -6,9 +6,9 @@ import (
 	campaigntypes "github.com/tendermint/spn/x/campaign/types"
 	launchtypes "github.com/tendermint/spn/x/launch/types"
 	profiletypes "github.com/tendermint/spn/x/profile/types"
+	"github.com/tendermint/starport/starport/pkg/cosmoserror"
 	"github.com/tendermint/starport/starport/pkg/cosmosutil"
 	"github.com/tendermint/starport/starport/pkg/events"
-	"github.com/tendermint/starport/starport/pkg/sdkerror"
 	"github.com/tendermint/starport/starport/services/network/networkchain"
 )
 
@@ -85,8 +85,8 @@ func (n Network) Publish(ctx context.Context, c Chain, options ...PublishOption)
 		CoordinatorByAddress(ctx, &profiletypes.QueryGetCoordinatorByAddressRequest{
 			Address: coordinatorAddress,
 		})
-	err = sdkerror.Unwrap(err)
-	if err == sdkerror.ErrInvalidRequest {
+	err = cosmoserror.Unwrap(err)
+	if err == cosmoserror.ErrInvalidRequest {
 		msgCreateCoordinator := profiletypes.NewMsgCreateCoordinator(
 			coordinatorAddress,
 			"",
@@ -107,7 +107,7 @@ func (n Network) Publish(ctx context.Context, c Chain, options ...PublishOption)
 				CampaignID: o.campaignID,
 			})
 		if err != nil {
-			return 0, 0, sdkerror.Unwrap(err)
+			return 0, 0, cosmoserror.Unwrap(err)
 		}
 	} else {
 		msgCreateCampaign := campaigntypes.NewMsgCreateCampaign(
@@ -117,12 +117,12 @@ func (n Network) Publish(ctx context.Context, c Chain, options ...PublishOption)
 		)
 		res, err := n.cosmos.BroadcastTx(n.account.Name, msgCreateCampaign)
 		if err != nil {
-			return 0, 0, sdkerror.Unwrap(err)
+			return 0, 0, cosmoserror.Unwrap(err)
 		}
 
 		var createCampaignRes campaigntypes.MsgCreateCampaignResponse
 		if err := res.Decode(&createCampaignRes); err != nil {
-			return 0, 0, sdkerror.Unwrap(err)
+			return 0, 0, cosmoserror.Unwrap(err)
 		}
 		campaignID = createCampaignRes.CampaignID
 	}
@@ -139,12 +139,12 @@ func (n Network) Publish(ctx context.Context, c Chain, options ...PublishOption)
 	)
 	res, err := n.cosmos.BroadcastTx(n.account.Name, msgCreateChain)
 	if err != nil {
-		return 0, 0, sdkerror.Unwrap(err)
+		return 0, 0, cosmoserror.Unwrap(err)
 	}
 
 	var createChainRes launchtypes.MsgCreateChainResponse
 	if err := res.Decode(&createChainRes); err != nil {
-		return 0, 0, sdkerror.Unwrap(err)
+		return 0, 0, cosmoserror.Unwrap(err)
 	}
 
 	return createChainRes.LaunchID, campaignID, nil
